+++ lib/bolt/inventory/target.rb
@@ -4,7 +4,7 @@ module Bolt
   class Inventory
     # This class represents the active state of a target within the inventory.
     class Target
-      attr_reader :name, :uri, :safe_name
+      attr_reader :name, :uri, :safe_name, :target_alias
 
       def initialize(target_data, inventory)
         unless target_data['name'] || target_data['uri']
@@ -34,7 +34,9 @@ module Bolt
         @features = target_data['features'] || Set.new
         @options = target_data['options'] || {}
         @plugin_hooks = target_data['plugin_hooks'] || {}
-        @target_alias = target_data['target_alias'] || []
+        # When alias is specified in a plan, the key will be `target_alias`, when
+        # alias is specified in inventory the key will be `alias`.
+        @target_alias = target_data['target_alias'] || target_data['alias'] || []
 
         @inventory = inventory
 
@@ -42,7 +44,6 @@ module Bolt
       end
 
       def vars
-        # XXX Return vars from the cache
         group_cache['vars'].merge(@vars)
       end
 
@@ -55,7 +56,6 @@ module Bolt
       # rubocop:enable Naming/AccessorMethodName
 
       def facts
-        # XXX Return facts from the cache
         Bolt::Util.deep_merge(group_cache['facts'], @facts)
       end