+++ spec/bolt/inventory/group2_spec.rb
@@ -53,6 +53,7 @@ describe Bolt::Inventory::Group2 do
                                                     'name' => nil,
                                                     'plugin_hooks' => {},
                                                     'uri' => 'target1',
+                                                    'alias' => nil,
                                                     'facts' => {},
                                                     'features' => Set.new,
                                                     'groups' => [])
@@ -612,7 +613,7 @@ describe Bolt::Inventory::Group2 do
         }
       end
 
-      it { expect { group.validate }.to raise_error(/Node name target1 conflicts with alias of the same name/) }
+      it { expect { group.validate }.to raise_error(/Target name target1 conflicts with alias of the same name/) }
     end
 
     context 'conflict with a later target name' do
@@ -626,7 +627,7 @@ describe Bolt::Inventory::Group2 do
         }
       end
 
-      it { expect { group.validate }.to raise_error(/Node name target2 conflicts with alias of the same name/) }
+      it { expect { group.validate }.to raise_error(/Target name target2 conflicts with alias of the same name/) }
     end
 
     context 'conflict with its own target name' do
@@ -644,7 +645,7 @@ describe Bolt::Inventory::Group2 do
       end
 
       it 'raises an error' do
-        expect { group.validate }.to raise_error(/Node name target1 conflicts with alias of the same name/)
+        expect { group.validate }.to raise_error(/Target name target1 conflicts with alias of the same name/)
       end
     end